Enable rotation of camera.
Added rotation state to mouse. getMove now returns new
camera direction or move of possition depending if left ctrl is
pressed.
In display we check for left control press and change state of
the mouse and activate/deactivate rotation.
Also we enable core profile of opengl so we use hdpi display
where possible.