Added zooming to mouse.
We can now zoom in application by using mousewheel.
Zooming is done by moving the camera in the forward direction
by some amount.
Refactored mouse class so it is much easier to reason about
and use it. Active state was renamed to button pressed meaning
mouse button is pressed down. We use this for moving and
rotation. Event is now copied to mouse class as it gets modified
latter on.
We reset mouse event if last event was mousewheel as we can
not detect when did we stop scrolling.